Move ServiceContext Provider to root
