+++ fusion-react/src/context.js
@@ -6,8 +6,6 @@
  * @flow
  */
 import * as React from 'react';
-import {createPlugin} from 'fusion-core';
-import type FusionApp, {FusionPlugin, Middleware} from 'fusion-core';
 import type {Element} from 'react';
 
 export const FusionContext = React.createContext<any>({});
@@ -44,35 +42,6 @@ export function ServiceConsumer<TService>({
       }}
     </ServiceContext.Consumer>
   );
-}
-
-export function serviceContextPlugin(app: FusionApp): FusionPlugin<void, void> {
-  function getService(token) {
-    const provides = app.getService(token);
-    const isRequiredToken = Boolean(token.optional);
-    if (typeof provides === 'undefined' && isRequiredToken) {
-      throw new Error(
-        `Token ${
-          token.name
-        } not registered or registered plugin does not provide a service. To use an optional plugin, use \`Token.optional\`.`
-      );
-    }
-    return provides;
-  }
-  return createPlugin({
-    middleware(): Middleware {
-      return (ctx, next) => {
-        ctx.element = ctx.element && (
-          <FusionContext.Provider value={ctx}>
-            <ServiceContext.Provider value={getService}>
-              {ctx.element}
-            </ServiceContext.Provider>
-          </FusionContext.Provider>
-        );
-        return next();
-      };
-    },
-  });
 }
 
 type Dependencies = {[string]: ReturnsType<mixed>};
